+++ CHANGELOG.rst
@@ -8,6 +8,10 @@ Changelog
 
 * Add Django 4.2+ fixer to combine deprecated ``DEFAULT_FILE_STORAGE`` and ``STATICFILES_STORAGE`` settings into the new ``STORAGES`` setting.
 
+* Add Django 4.2+ fixer to rewrite HTTP headers passed to test ``Client`` and ``RequestFactory``.
+
+  Thanks to David Wobrock in `PR #305 <https://github.com/adamchainz/django-upgrade/pull/305>`__.
+
 * Add Django 4.0 fixer to rewrite ``django.contrib.admin.utils.lookup_needs_distinct`` -> ``lookup_spawns_duplicates``.
 
   Thanks to Bruno Alla in `PR #313 <https://github.com/adamchainz/django-upgrade/pull/313>`__.